Change Int's to Doubles
Change Int's to Doubles on Player methods to match the Enemy Class.

+++ src/main/java/game/rogue/PlayerClass.java
@@ -1,12 +1,12 @@
 package game.rogue;
 
 public abstract class PlayerClass {
-    private int baseHealth;
-    private int baseMana;
+    private double baseHealth;
+    private double baseMana;
 
-    public abstract int getMaxHitPoints(int level);
+    public abstract double getMaxHitPoints(int level);
 
-    public abstract int getMaxMana(int level);
+    public abstract double getMaxMana(int level);
 
     public boolean canEquipCloth(){
         return false;
